// server/src/models/filterOperator.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * IDHI API
 * OpenAPI spec version: 1.0.0
 */

export type FilterOperator =
  (typeof FilterOperator)[keyof typeof FilterOperator]

export const FilterOperator = {
  eq: 'eq',
  ne: 'ne',
  gt: 'gt',
  gte: 'gte',
  lt: 'lt',
  lte: 'lte',
  in: 'in',
  nin: 'nin',
  exists: 'exists',
} as const